refactor: Move RegisterUIState to uiState package
