+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/register/createPassword/CreatePasswordUIEffect.kt
@@ -1,6 +1,6 @@
 package net.thechance.mena.identity.presentation.screen.register.createPassword
 
-import net.thechance.mena.identity.presentation.screen.register.shared.RegisterUIState
+import net.thechance.mena.identity.presentation.screen.register.shared.uiState.RegisterUIState
 
 sealed interface CreatePasswordUIEffect {
     data class NavigateToDatePicker(val registerUIState: RegisterUIState) : CreatePasswordUIEffect
