+++ identity_presentation/src/commonMain/kotlin/net/thechance/mena/identity/presentation/screen/register/phoneEntry/RegisterPhoneEntryUIEffect.kt
@@ -1,6 +1,6 @@
 package net.thechance.mena.identity.presentation.screen.register.phoneEntry
 
-import net.thechance.mena.identity.presentation.screen.register.shared.RegisterUIState
+import net.thechance.mena.identity.presentation.screen.register.shared.uiState.RegisterUIState
 
 sealed interface RegisterPhoneEntryUIEffect {
     data class NavigateToOTP(val registerUIState: RegisterUIState) : RegisterPhoneEntryUIEffect
